requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
python-dotenv>=1.0.0

numpy>=2.1.0
scipy>=1.13.0
soundfile>=0.12.1
librosa>=0.10.2

torch>=2.1.2; python_version < "3.13"
torchaudio>=2.1.2; python_version < "3.13"
transformers>=4.36.2; python_version < "3.13"
